--- src/main/java/org/matsim/run/OpenBerlinScenarioSEVC.java
@@ -1,282 +1,0 @@
-package org.matsim.run;
-
-import java.util.LinkedList;
-import java.util.List;
-
-import org.matsim.analysis.QsimTimingModule;
-import org.matsim.analysis.personMoney.PersonMoneyEventsAnalysisModule;
-import org.matsim.api.core.v01.Scenario;
-import org.matsim.api.core.v01.TransportMode;
-import org.matsim.api.core.v01.population.Person;
-import org.matsim.application.MATSimApplication;
-import org.matsim.application.options.SampleOptions;
-import org.matsim.contrib.bicycle.BicycleConfigGroup;
-import org.matsim.contrib.bicycle.BicycleLinkSpeedCalculator;
-import org.matsim.contrib.bicycle.BicycleLinkSpeedCalculatorDefaultImpl;
-import org.matsim.contrib.bicycle.BicycleParams;
-import org.matsim.contrib.bicycle.BicycleParamsDefaultImpl;
-import org.matsim.contrib.bicycle.BicycleTravelTime;
-import org.matsim.contrib.emissions.HbefaRoadTypeMapping;
-import org.matsim.contrib.emissions.OsmHbefaMapping;
-import org.matsim.contrib.emissions.utils.EmissionsConfigGroup;
-import org.matsim.contrib.ev.EvConfigGroup;
-import org.matsim.contrib.ev.EvModule;
-import org.matsim.contrib.ev.infrastructure.ChargingInfrastructureSpecification;
-import org.matsim.contrib.ev.strategic.StrategicChargingUtils;
-import org.matsim.contrib.vsp.scoring.RideScoringParamsFromCarParams;
-import org.matsim.core.config.Config;
-import org.matsim.core.config.ConfigUtils;
-import org.matsim.core.config.groups.ReplanningConfigGroup;
-import org.matsim.core.config.groups.VspExperimentalConfigGroup;
-import org.matsim.core.controler.AbstractModule;
-import org.matsim.core.controler.Controler;
-import org.matsim.core.population.PopulationUtils;
-import org.matsim.core.replanning.strategies.DefaultPlanStrategiesModule;
-import org.matsim.core.replanning.strategies.DefaultPlanStrategiesModule.DefaultSelector;
-import org.matsim.core.router.costcalculators.OnlyTimeDependentTravelDisutilityFactory;
-import org.matsim.core.router.costcalculators.TravelDisutilityFactory;
-import org.matsim.core.router.util.TravelTime;
-import org.matsim.core.scoring.functions.ScoringParametersForPerson;
-import org.matsim.run.StrategicChargingConfigurator.Settings;
-import org.matsim.run.scoring.AdvancedScoringConfigGroup;
-import org.matsim.run.scoring.AdvancedScoringModule;
-import org.matsim.simwrapper.SimWrapperConfigGroup;
-import org.matsim.simwrapper.SimWrapperModule;
-
-import com.google.inject.Key;
-import com.google.inject.name.Names;
-
-import picocli.CommandLine;
-import playground.vsp.scoring.IncomeDependentUtilityOfMoneyPersonScoringParameters;
-
-@CommandLine.Command(header = ":: Open Berlin Scenario ::", version = OpenBerlinScenario.VERSION, mixinStandardHelpOptions = true, showDefaultValues = true)
-public class OpenBerlinScenarioSEVC extends MATSimApplication {
-
-	public static final String VERSION = "6.4";
-	public static final String CRS = "EPSG:25832";
-
-	//	To decrypt hbefa input files set MATSIM_DECRYPTION_PASSWORD as environment variable. ask VSP for access.
-	private static final String HBEFA_2020_PATH = "https://svn.vsp.tu-berlin.de/repos/public-svn/3507bb3997e5657ab9da76dbedbb13c9b5991d3e/0e73947443d68f95202b71a156b337f7f71604ae/";
-	private static final String HBEFA_FILE_COLD_DETAILED = HBEFA_2020_PATH + "82t7b02rc0rji2kmsahfwp933u2rfjlkhfpi2u9r20.enc";
-	private static final String HBEFA_FILE_WARM_DETAILED = HBEFA_2020_PATH + "944637571c833ddcf1d0dfcccb59838509f397e6.enc";
-	private static final String HBEFA_FILE_COLD_AVERAGE = HBEFA_2020_PATH + "r9230ru2n209r30u2fn0c9rn20n2rujkhkjhoewt84202.enc" ;
-	private static final String HBEFA_FILE_WARM_AVERAGE = HBEFA_2020_PATH + "7eff8f308633df1b8ac4d06d05180dd0c5fdf577.enc";
-
-	@CommandLine.Mixin
-	private final SampleOptions sample = new SampleOptions(10, 25, 3, 1);
-
-	@CommandLine.Option(names = "--plan-selector",
-		description = "Plan selector to use.",
-		defaultValue = DefaultPlanStrategiesModule.DefaultSelector.ChangeExpBeta)
-	private String planSelector;
-
-	public OpenBerlinScenarioSEVC() {
-		super(String.format("input/v%s/berlin-v%s.config.xml", VERSION, VERSION));
-	}
-
-	// SEVC
-	private final Settings sevcSettings = new Settings();
-	private final StrategicChargingConfigurator sevcConfigurator = new StrategicChargingConfigurator(sevcSettings);
-	private ChargingInfrastructureSpecification infrastructureSpecification;
-
-	public static void main(String[] args) {
-		MATSimApplication.run(OpenBerlinScenarioSEVC.class, args);
-	}
-
-	@Override
-	protected Config prepareConfig(Config config) {
-
-		SimWrapperConfigGroup sw = ConfigUtils.addOrGetModule(config, SimWrapperConfigGroup.class);
-
-		if (sample.isSet()) {
-			double sampleSize = sample.getSample();
-
-			config.qsim().setFlowCapFactor(sampleSize);
-			config.qsim().setStorageCapFactor(sampleSize);
-
-			// Counts can be scaled with sample size
-			config.counts().setCountsScaleFactor(sampleSize);
-			sw.setSampleSize(sampleSize);
-
-			config.controller().setRunId(sample.adjustName(config.controller().getRunId()));
-			config.controller().setOutputDirectory(sample.adjustName(config.controller().getOutputDirectory()));
-			config.plans().setInputFile(sample.adjustName(config.plans().getInputFile()));
-		}
-
-		config.qsim().setUsingTravelTimeCheckInTeleportation(true);
-
-		// overwrite ride scoring params with values derived from car
-		RideScoringParamsFromCarParams.setRideScoringParamsBasedOnCarParams(config.scoring(), 1.0);
-		Activities.addScoringParams(config, true);
-
-		// Required for all calibration strategies
-		for (String subpopulation : List.of("person", "freight", "goodsTraffic", "commercialPersonTraffic", "commercialPersonTraffic_service")) {
-			config.replanning().addStrategySettings(
-				new ReplanningConfigGroup.StrategySettings()
-					.setStrategyName(planSelector)
-					.setWeight(1.0)
-					.setSubpopulation(subpopulation)
-			);
-
-			config.replanning().addStrategySettings(
-				new ReplanningConfigGroup.StrategySettings()
-					.setStrategyName(DefaultPlanStrategiesModule.DefaultStrategy.ReRoute)
-					.setWeight(0.15)
-					.setSubpopulation(subpopulation)
-			);
-		}
-
-		config.replanning().addStrategySettings(
-			new ReplanningConfigGroup.StrategySettings()
-				.setStrategyName(DefaultPlanStrategiesModule.DefaultStrategy.TimeAllocationMutator)
-				.setWeight(0.15)
-				.setSubpopulation("person")
-		);
-
-		config.replanning().addStrategySettings(
-			new ReplanningConfigGroup.StrategySettings()
-				.setStrategyName(DefaultPlanStrategiesModule.DefaultStrategy.SubtourModeChoice)
-				.setWeight(0.15)
-				.setSubpopulation("person")
-		);
-
-		// Need to switch to warning for best score
-		if (planSelector.equals(DefaultPlanStrategiesModule.DefaultSelector.BestScore)) {
-			config.vspExperimental().setVspDefaultsCheckingLevel(VspExperimentalConfigGroup.VspDefaultsCheckingLevel.warn);
-		}
-
-		// Bicycle config must be present
-		ConfigUtils.addOrGetModule(config, BicycleConfigGroup.class);
-
-		// Add emissions configuration
-		EmissionsConfigGroup eConfig = ConfigUtils.addOrGetModule(config, EmissionsConfigGroup.class);
-		eConfig.setDetailedColdEmissionFactorsFile(HBEFA_FILE_COLD_DETAILED);
-		eConfig.setDetailedWarmEmissionFactorsFile(HBEFA_FILE_WARM_DETAILED);
-		eConfig.setAverageColdEmissionFactorsFile(HBEFA_FILE_COLD_AVERAGE);
-		eConfig.setAverageWarmEmissionFactorsFile(HBEFA_FILE_WARM_AVERAGE);
-		eConfig.setHbefaTableConsistencyCheckingLevel(EmissionsConfigGroup.HbefaTableConsistencyCheckingLevel.consistent);
-		eConfig.setDetailedVsAverageLookupBehavior(EmissionsConfigGroup.DetailedVsAverageLookupBehavior.tryDetailedThenTechnologyAverageThenAverageTable);
-		eConfig.setEmissionsComputationMethod(EmissionsConfigGroup.EmissionsComputationMethod.StopAndGoFraction);
-
-		EvConfigGroup evConfig = new EvConfigGroup();
-		evConfig.setChargersFile("injected");
-		config.addModule(evConfig);
-
-		// avoid aborting
-		config.vspExperimental().setVspDefaultsCheckingLevel(VspExperimentalConfigGroup.VspDefaultsCheckingLevel.warn);
-
-		return config;
-	}
-
-	@Override
-	protected void prepareScenario(Scenario scenario) {
-
-		// add hbefa link attributes.
-		HbefaRoadTypeMapping roadTypeMapping = OsmHbefaMapping.build();
-		roadTypeMapping.addHbefaMappings(scenario.getNetwork());
-
-		// back up non-person agents
-		List<Person> backup = new LinkedList<>();
-		for (Person person : scenario.getPopulation().getPersons().values()) {
-			if (!PopulationUtils.getSubpopulation(person).equals("person")) {
-				backup.add(person);
-			}
-		}
-
-		for (Person person : backup) {
-			scenario.getPopulation().removePerson(person.getId());
-		}
-
-		// SEVC: We only work with persons
-		infrastructureSpecification = sevcConfigurator.apply(scenario);
-
-		// add them back
-		backup.forEach(scenario.getPopulation()::addPerson);
-
-		for (var strategy : scenario.getConfig().replanning().getStrategySettings()) {
-			strategy.setSubpopulation("person");
-		}
-
-		for (String subpopulation : List.of("freight", "goodsTraffic", "commercialPersonTraffic", "commercialPersonTraffic_service")) {
-			scenario.getConfig().replanning().addStrategySettings(
-				new ReplanningConfigGroup.StrategySettings()
-					.setStrategyName(DefaultSelector.KeepLastSelected)
-					.setWeight(1.0)
-					.setSubpopulation(subpopulation)
-			);
-		}
-	}
-
-	@Override
-	protected void prepareControler(Controler controler) {
-
-		controler.addOverridingModule(new SimWrapperModule());
-
-		controler.addOverridingModule(new TravelTimeBinding());
-
-		controler.addOverridingModule(new QsimTimingModule());
-
-		// AdvancedScoring is specific to matsim-berlin!
-		if (ConfigUtils.hasModule(controler.getConfig(), AdvancedScoringConfigGroup.class)) {
-			controler.addOverridingModule(new AdvancedScoringModule());
-			controler.getConfig().scoring().setExplainScores(true);
-		} else {
-			// if the above config group is not present we still need income dependent scoring
-			// this implementation also allows for person specific asc
-			controler.addOverridingModule(new AbstractModule() {
-				@Override
-				public void install() {
-					bind(ScoringParametersForPerson.class).to(IncomeDependentUtilityOfMoneyPersonScoringParameters.class).asEagerSingleton();
-				}
-			});
-		}
-		controler.addOverridingModule(new PersonMoneyEventsAnalysisModule());
-
-		// SEVC
-		controler.addOverridingModule(new EvModule());
-
-		controler.addOverridingModule(new AbstractModule() {
-			@Override
-			public void install() {
-				bind(ChargingInfrastructureSpecification.class).toInstance(infrastructureSpecification);
-			}
-		});
-
-		StrategicChargingUtils.configureController(controler);
-	}
-
-	/**
-	 * Add travel time bindings for ride and freight modes, which are not actually network modes.
-	 */
-	public static final class TravelTimeBinding extends AbstractModule {
-
-		private final boolean carOnly;
-
-		public TravelTimeBinding() {
-			this.carOnly = false;
-		}
-
-		public TravelTimeBinding(boolean carOnly) {
-			this.carOnly = carOnly;
-		}
-
-		@Override
-		public void install() {
-			addTravelTimeBinding(TransportMode.ride).to(networkTravelTime());
-			addTravelDisutilityFactoryBinding(TransportMode.ride).to(carTravelDisutilityFactoryKey());
-
-			if (!carOnly) {
-				addTravelTimeBinding("freight").to(Key.get(TravelTime.class, Names.named(TransportMode.truck)));
-				addTravelDisutilityFactoryBinding("freight").to(Key.get(TravelDisutilityFactory.class, Names.named(TransportMode.truck)));
-
-
-				bind(BicycleLinkSpeedCalculator.class).to(BicycleLinkSpeedCalculatorDefaultImpl.class);
-				bind(BicycleParams.class).to(BicycleParamsDefaultImpl.class);
-
-				// Bike should use free speed travel time
-				addTravelTimeBinding(TransportMode.bike).to(BicycleTravelTime.class);
-				addTravelDisutilityFactoryBinding(TransportMode.bike).to(OnlyTimeDependentTravelDisutilityFactory.class);
-			}
-		}
-	}
-}
